>  run
qile
Linux debian 4.19.0-5-amd64 #1 SMP Debian 4.19.37-5+deb10u2 (2019-08-08) x86_64 GNU/Linux
 21:55:25 up  6:36,  1 user,  load average: 0.27, 0.22, 0.18
261a859f1f09b0ce4f9bf0921ad8699b288b49fe

+++ nexus-am/am/src/nemu-common/nemu-input.c
@@ -8,7 +8,7 @@ size_t __am_input_read(uintptr_t reg, void *buf, size_t size) {
   switch (reg) {
     case _DEVREG_INPUT_KBD: {
       _DEV_INPUT_KBD_t *kbd = (_DEV_INPUT_KBD_t *)buf;
-      int fetch = inl(0x60);
+      unsigned long long fetch = inl(0x60);
       kbd->keycode = fetch;
       if(fetch!=_KEY_NONE){
 	      kbd->keydown=1-kbd->keydown;
